@@ -27,7 +27,7 @@ classifiers = [
             "Programming Language :: Python :: 3",
             "Topic :: Scientific/Engineering :: Visualization",
 ]
-requires-python = ">3.6"
+requires-python = ">=3.9"
 dependencies = [
                "numpy",
                "scipy", 
